pyproject.toml
[project]
name = "vai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<=3.12"
dependencies = [
    "accelerate>=1.7.0",
    "bitsandbytes>=0.46.0",
    "black>=25.1.0",
    "ctranslate2>=4.6.0",
    "datasets>=3.6.0",
    "fastapi>=0.115.12",
    "faster-whisper>=1.1.1",
    "gradio>=4.44.1",
    "langchain>=0.3.25",
    "numpy>=1.22,<2.0",
    "peft>=0.15.2",
    "piper-tts>=1.2.0",
    "psycopg2>=2.9.10",
    "requests>=2.32.3",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "torch>=2.7.0",
    "transformers>=4.51.3",
    "uvicorn>=0.34.2",
]
